+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+
+- Add Example project
+- Fix Pub Sample Image
+
 ## 0.0.2
 
 - Update Docs